+++ src/ui/svelte/window_manager/utils.ts
@@ -24,4 +24,19 @@ export class NodeUtils {
         return node.children.some((child) => NodeUtils.contains(child, searchingWindow));
     }
   }
+
+  static some(node: WmNode, predicate: (window: number) => boolean): boolean {
+    switch (node.type) {
+      case WmNodeKind.Leaf:
+        return predicate(node.active!);
+      case WmNodeKind.Stack:
+        if (node.active) {
+          return predicate(node.active);
+        }
+        return false;
+      case WmNodeKind.Horizontal:
+      case WmNodeKind.Vertical:
+        return node.children.some((child) => NodeUtils.some(child, predicate));
+    }
+  }
 }