+++ changelog.md
@@ -36,6 +36,7 @@
 - fix stale battery percentage on sleep wake up.
 - click on keyboard layout no changing the layout on the system.
 - window labels not being shown on dock.
+- tiling window manager not pausing on maximized/fullscreen apps.
 
 ## [2.4.11]